Utilizing the  data
